Make sendAssets modal to add action into global store and navigate into proposal with actions view, instead of creating proposal right away
