+++ src/i18n/locales/en/modals.json
@@ -11,16 +11,13 @@
   "pendingDelegateVote": "Delegating vote",
   "failedDelegateVote": "Delegation failed",
   "successDelegateVote": "Delegation successful",
-  "sendAssetsTitle": "Send Assets",
+  "sendAssetsTitle": "Request Payment",
   "selectLabel": "Select Asset",
   "selectSublabel": "Available balance {{balance}}",
   "amountLabel": "How much?",
   "destinationLabel": "Where would you like to send this?",
   "destinationSublabel": "Ethereum wallet address or ENS only",
   "submitProposal": "Submit Proposal",
-  "sendAssetsPendingToastMessage": "Creating proposal to send assets",
-  "sendAssetsSuccessToastMessage": "Send assets proposal successfully created",
-  "sendAssetsFailureToastMessage": "Send assets proposal failed",
   "confirmUrlTitle": "This link will take you to",
   "confirmAction": "Do you want to continue?",
   "modalContinue": "Continue",
